Fix OSX build

+++ source/time/timestamp.cpp
@@ -124,7 +124,7 @@ uint64_t Timestamp::utc()
     result = mach_port_deallocate(task_port, cclock);
     if (result != KERN_SUCCESS)
         throwex SystemException("Cannot release the current host clock service!");
-    uint64_t result1 = (mts.tv_sec * 1000 * 1000 * 1000) + mts.tv_nsec;
+    uint64_t result1 = ((uint64_t)mts.tv_sec * 1000 * 1000 * 1000) + mts.tv_nsec;
     std::cerr << "Timestamp: " << result1 << std::endl;
     return result1;
 #elif defined(unix) || defined(__unix) || defined(__unix__)
